Add all rpm testing logic to exptest driver for BL
